omnilatam/apps/order/api/serializers/order.py
"""Order Serializer."""

# Django REST Framework
from rest_framework import serializers

# Models
from omnilatam.apps.user.models import Profile, User
from omnilatam.apps.order.models.order import OrderProduct, Order, ProductsOrdered
from omnilatam.apps.order.models.product import Category, Product, ProductRelated